requirements.txt
Cython==0.29.15
dm-tree==0.1.2
# git+https://github.com/chrodan/joblib.git
ipdb==0.13.2
matplotlib==3.2.1
numpy==1.18.2
ray[tune]==0.8.4
scikit-learn==0.22.2.post1
tensorboard  # ==2.1.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorboardX==2.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-probability==0.9.0
